Clean up jest-config/jest-runtime/jest-cli.
