+++ packages/jest-resolve/src/index.js
@@ -12,11 +12,11 @@
 
 import type {
   FileMetaData,
-  HasteMap,
   ModuleMap,
   HTypeValue,
 } from 'types/HasteMap';
-import type {Path} from 'types/Config';
+import type HasteMap from '../../jest-haste-map';
+import type {Config, Path} from 'types/Config';
 
 const H = require('jest-haste-map').H;
 
@@ -32,11 +32,11 @@ export type MockedModuleContext = {
 };
 
 type ResolverConfig = {
-  defaultPlatform: string,
+  defaultPlatform: ?string,
   extensions: Array<string>,
   hasCoreModules: boolean,
   moduleDirectories: string | Array<string>,
-  moduleNameMapper: {[key: string]: RegExp},
+  moduleNameMapper: ?{[key: string]: RegExp},
   modulePaths: Array<Path>,
   platforms?: Array<string>,
 };
@@ -54,6 +54,17 @@ const NATIVE_PLATFORM = 'native';
 
 const nodePaths =
   (process.env.NODE_PATH ? process.env.NODE_PATH.split(path.delimiter) : null);
+
+const getModuleNameMapper = (config: Config) => {
+  if (config.moduleNameMapper.length) {
+    const moduleNameMapper = Object.create(null);
+    config.moduleNameMapper.forEach(
+      map => moduleNameMapper[map[1]] = new RegExp(map[0])
+    );
+    return moduleNameMapper;
+  }
+  return null;
+};
 
 class Resolver {
   _options: ResolverConfig;
@@ -78,6 +89,21 @@ class Resolver {
     this._moduleMap = moduleMap;
     this._moduleNameCache = Object.create(null);
     this._modulePathCache = Object.create(null);
+  }
+
+  static createResolver(
+    config: Config,
+    moduleMap: MockedModuleContext,
+  ): Resolver {
+    return new Resolver(moduleMap, {
+      defaultPlatform: config.haste.defaultPlatform,
+      extensions: config.moduleFileExtensions.map(extension => '.' + extension),
+      hasCoreModules: true,
+      moduleDirectories: config.moduleDirectories,
+      moduleNameMapper: getModuleNameMapper(config),
+      modulePaths: config.modulePaths,
+      platforms: config.haste.platforms,
+    });
   }
 
   static findNodeModule(path: Path, options: FindNodeModuleConfig): ?Path {
